@@ -1,5 +1,3 @@
-// ignore_for_file: avoid_print
-
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:on_audio_query/on_audio_query.dart';
@@ -31,36 +29,29 @@ class _LocalMusicArtistState extends State<LocalMusicArtist> {
   Widget build(BuildContext context) {
     return Consumer2<ListLongPress, AudioOptionSelected>(
       builder: (context, llp, aos, child) {
-        print(
-          "localmusic index 下artist中 ${llp.isPlaylistLongPress} ${llp.selectedPlaylistList.length} ${llp.localMusicAppBarSearchInput}",
-        );
+        /// 构建歌手列表，需要监测appbar中条件查询变化、排序选项变化，并及时更新显示符合条件的列表
 
-        /// 如果是在播放列表中对某音频进行了长按，则在此处显示一些功能按钮
-        ///   暂时有：查看信息、从当前列表移除、三个点（添加到播放列表、添加到队列(这个暂不实现)、全选等）
-        /// 如果是默认显示的，应该有：排序、搜索、三个点（展开其他功能）
+        // 如果是主页上歌单的条件查询
+        if (llp.localMusicAppBarSearchInput != null) {
+          futureHandler = _audioQuery.queryWithFilters(
+            llp.localMusicAppBarSearchInput!,
+            WithFiltersType.ARTISTS,
+          );
+        } else {
+          // 如果等于null，说明是初始化，或者关闭了查询按钮，歌单要重新查询所有。
+          // 此外，在排序时，也是直接获取到对应的排序类别和用于排序的关键字；如果是条件查询，则不对结果排序了(也不知道怎么排)
+          futureHandler = _audioQuery.queryArtists(
+            sortType: aos.artistSortType,
+            orderType: aos.orderType,
+          );
+        }
+
         return _buildList(context, llp, aos);
       },
     );
   }
 
   _buildList(BuildContext context, ListLongPress llp, AudioOptionSelected aos) {
-    // 如果是主页上歌单的条件查询
-    if (llp.localMusicAppBarSearchInput != null) {
-      print("【歌手】执行了条件查询的逻辑");
-      futureHandler = _audioQuery.queryWithFilters(
-        llp.localMusicAppBarSearchInput!,
-        WithFiltersType.ARTISTS,
-      );
-    } else {
-      // 如果等于null，说明是初始化，或者关闭了查询按钮，歌单要重新查询所有。
-      // 此外，在排序时，也是直接获取到对应的排序类别和用于排序的关键字；如果是条件查询，则不对结果排序了(也不知道怎么排)
-      print("执行了【歌单初始化】或者【关闭】条件查询的逻辑 ${aos.artistSortType}");
-      futureHandler = _audioQuery.queryArtists(
-        sortType: aos.artistSortType,
-        orderType: aos.orderType,
-      );
-    }
-
     return FutureBuilder<List<dynamic>>(
       future: futureHandler,
       builder: (context, item) {
@@ -74,20 +65,22 @@ class _LocalMusicArtistState extends State<LocalMusicArtist> {
         }
         // 'Library' is empty.
         if (item.data!.isEmpty) {
-          return const Center(child: Text("暂无歌手!"));
+          return const Center(child: Text("暂无歌手"));
         }
 
-        // 得到查询的歌手列表
+        // 得到查询的歌手列表(条件查询结果是动态类型，但可以转为艺术家模型对象)
         List<ArtistModel> artists = item.data! is List<ArtistModel>
             ? item.data! as List<ArtistModel>
             : item.data!.toArtistModel();
 
         return ListView.builder(
           itemCount: artists.length,
-          itemExtent: 80.sp, // 每个item内部组件的高度(因为下面leading的高度有问题，这里暂时上下有点间距)
+          // 每个item内部组件的高度(因为下面leading的高度有问题，这里暂时上下有点间距)
+          itemExtent: 80.sp,
           itemBuilder: (context, index) {
             return ListTile(
-              minLeadingWidth: 100.sp, // 左侧缩略图标的最小宽度
+              // 左侧缩略图标的最小宽度
+              minLeadingWidth: 100.sp,
               title: Text(artists[index].artist),
               subtitle: Text(
                 "${artists[index].numberOfTracks.toString()} 张专辑 ${artists[index].numberOfTracks.toString()} 首歌曲",
@@ -100,16 +93,20 @@ class _LocalMusicArtistState extends State<LocalMusicArtist> {
                 type: ArtworkType.ARTIST,
                 artworkBorder: const BorderRadius.all(Radius.zero), // 缩略图不显示圆角
                 artworkWidth: 100.sp, // 默认是50*50的大小
-                artworkHeight: 100.sp, // 这个高度显示不太对，实测始终是56，原因不明
+                // 这个高度显示不太对，实测始终是56，原因不明
+                // 2024-01-09 因为listtile的限制，最高就56
+                artworkHeight: 100.sp,
                 artworkFit: BoxFit.cover,
                 keepOldArtwork: true, // 在生命周期内使用旧的缩略图
-                nullArtworkWidget: const SizedBox.shrink(),
+                // nullArtworkWidget: const SizedBox.shrink(),
+                // 没有缩略图时使用占位图
+                nullArtworkWidget: SizedBox(
+                  width: 100.sp,
+                  child: Image.asset(placeholderImageUrl, fit: BoxFit.fitWidth),
+                  // child: Icon(Icons.image, size: 50.sp),
+                ),
               ),
               onTap: () {
-                print(
-                  '指定歌手 ${artists[index].artist}  was tapped! id Is ${artists[index].id}',
-                );
-
                 Navigator.of(context).push(
                   MaterialPageRoute(
                     // 在选中指定歌手点击后，进入音频列表，同时监控是否有对音频长按
