@@ -1,5 +1,3 @@
-// ignore_for_file: avoid_print
-
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:on_audio_query/on_audio_query.dart';
@@ -14,10 +12,7 @@ import '../../services/service_locator.dart';
 import 'nested_pages/audio_list_detail.dart';
 
 class LocalMusicPlaylist extends StatefulWidget {
-  const LocalMusicPlaylist({super.key, this.queryInputted});
-
-  // 2023-04-14 如果是主页面“全部”tab的查询结果，可能把查询条件传过来。如果有，则用它；没有，才使用initFuture。
-  final String? queryInputted;
+  const LocalMusicPlaylist({super.key});
 
   @override
   State<LocalMusicPlaylist> createState() => _LocalMusicPlaylistState();
@@ -34,97 +29,44 @@ class _LocalMusicPlaylistState extends State<LocalMusicPlaylist> {
   List<PlaylistModel> selectedPlaylists = [];
 
   @override
-  void initState() {
-    super.initState();
-
-    testData();
-  }
-
-  testData() async {
-    var list = await _audioQuery.queryPlaylists(
-      orderType: OrderType.DESC_OR_GREATER,
-      sortType: PlaylistSortType.PLAYLIST,
-    );
-    var list2 = await _audioQuery.queryPlaylists(
-      orderType: OrderType.ASC_OR_SMALLER,
-      sortType: PlaylistSortType.PLAYLIST,
-    );
-
-    var list3 = await _audioQuery.queryAudiosFrom(
-      AudiosFromType.PLAYLIST,
-      213023,
-      sortType: SongSortType.TITLE,
-      orderType: OrderType.ASC_OR_SMALLER,
-    );
-
-    print("888888888888888888888888888888888");
-    print(list);
-    print(list2);
-    print(list3);
-    print("888888888888888888888888888888888");
-  }
-
-  // 进入指定歌单查看音频之后返回，需要重新查询歌单列表，以防有音频的删除信息而歌单列表显示未更新的问题
-  initData() async {
-    print("在歌单列表的init中，传入的widget.queryInputted:${widget.queryInputted}|");
-
-    setState(() {
-      futureHandler = _audioQuery.queryPlaylists();
-    });
-  }
-
-  @override
   Widget build(BuildContext context) {
     return Consumer2<ListLongPress, AudioOptionSelected>(
       builder: (context, llp, aos, child) {
-        print(
-          "localmusic index 下playlist中 ${llp.isPlaylistLongPress} ${llp.selectedPlaylistList.length} ${llp.localMusicAppBarSearchInput}"
-          "${aos.orderType} ${aos.playlistSortType}",
-        );
+        /// 构建歌手列表，需要监测appbar中条件查询变化、排序选项变化、和专辑跟歌手没有的“长按状态变化”，并及时更新显示符合条件的列表
 
-        /// 如果是在播放列表中对某音频进行了长按，则在此处显示一些功能按钮
-        ///   暂时有：查看信息、从当前列表移除、三个点（添加到播放列表、添加到队列(这个暂不实现)、全选等）
-        /// 如果是默认显示的，应该有：排序、搜索、三个点（展开其他功能）
+        // 如果是上层使用provide取消了长按标志，这里得清空被选中的数组
+        // 但是默认就是false，是不是初始化的时候也会进来？
+        if (llp.isPlaylistLongPress == LongPressStats.NO) {
+          selectedPlaylists.length = 0;
+          // 取消歌单长按，有可能是删除了歌单，那么需要刷新一下歌单数据
+          futureHandler = _audioQuery.queryPlaylists();
+          // 取消完之后重新查询了歌单，则修改状态为初始化
+          llp.changeIsPlaylistLongPress(LongPressStats.INIT);
+        }
+
+        // 如果是主页上歌单的条件查询
+        if (llp.localMusicAppBarSearchInput != null) {
+          futureHandler = _audioQuery.queryWithFilters(
+            llp.localMusicAppBarSearchInput!,
+            WithFiltersType.PLAYLISTS,
+          );
+        } else {
+          // 如果等于null，说明是初始化，或者关闭了查询按钮，歌单要重新查询所有
+          // 此外，在对歌单排序时，也是直接获取到对应的排序类别和用于排序的关键字；如果是条件查询，则不对结果排序了(也不知道怎么排)
+          futureHandler = _audioQuery.queryPlaylists(
+            sortType: aos.playlistSortType,
+            orderType: aos.orderType,
+          );
+        }
+
+        /// 注意，和歌手与专辑列表查询的结果不同，这里歌单返回的数据比较乱，部分需要自己处理成符合条件的结构
         return _buildList(context, llp, aos);
       },
     );
-    // return _buildList(context);
   }
 
   _buildList(BuildContext context, ListLongPress llp, AudioOptionSelected aos) {
-    // 如果是上层使用provide取消了长按标志，这里得清空被选中的数组
-    // 但是默认就是false，是不是初始化的时候也会进来？
-    if (llp.isPlaylistLongPress == LongPressStats.NO) {
-      print("执行取消选择的歌单的逻辑");
-      selectedPlaylists.length = 0;
-      // 取消歌单长按，有可能是删除了歌单，那么需要刷新一下歌单数据
-      futureHandler = _audioQuery.queryPlaylists();
-      // 取消完之后重新查询了歌单，则修改状态为初始化
-      llp.changeIsPlaylistLongPress(LongPressStats.INIT);
-    }
-
-    // 如果是主页上歌单的条件查询
-    if (llp.localMusicAppBarSearchInput != null) {
-      print("执行了条件查询的逻辑");
-      futureHandler = _audioQuery.queryWithFilters(
-        llp.localMusicAppBarSearchInput!,
-        WithFiltersType.PLAYLISTS,
-      );
-    } else {
-      // 如果等于null，说明是初始化，或者关闭了查询按钮，歌单要重新查询所有
-      print(
-        "执行了【歌单初始化】、【关闭】、【排序】条件查询的逻辑"
-        "${aos.orderType} ${aos.playlistSortType}",
-      );
-      // 此外，在对歌单排序时，也是直接获取到对应的排序类别和用于排序的关键字；如果是条件查询，则不对结果排序了(也不知道怎么排)
-      futureHandler = _audioQuery.queryPlaylists(
-        sortType: aos.playlistSortType,
-        orderType: aos.orderType,
-      );
-    }
-
     // 如果切换了歌单的排序方式，则按照给出的排序方式重新查询
-
     return FutureBuilder<List<dynamic>>(
       future: futureHandler,
       builder: (context, item) {
@@ -144,22 +86,16 @@ class _LocalMusicPlaylistState extends State<LocalMusicPlaylist> {
         // 得到查询的歌单列表
         List<dynamic> playlists = item.data!;
 
-        print(
-          "在playlist.dart页面中得到的List<PlaylistModel> playlists ${llp.localMusicAppBarSearchInput} ${item.data!.runtimeType} ${item.data!}",
-        );
-
         return ListView.builder(
           itemCount: playlists.length,
           itemExtent: 80.sp, // 每个item内部组件的高度(因为下面leading的高度有问题，这里暂时上下有点间距)
           itemBuilder: (ctx, index) {
-            // PlaylistModel playlist = playlists[index];
-
             // 因为 queryWithFilters 查询 playlist的时候有bug，没有 numOfSongs 属性。所以转为PlaylistModel会报错
             // 所以关于playlist的取值，都转为map，然后用中括号获取id属性。
             // 然后使用 queryAudiosFrom() 方法，获取到缺少的 numOfSongs，补充到该map中。
             // 最后使用 PlaylistModel的构造函数，把map还原为PlaylistModel类型。
             var playlistMap = (playlists[index] is PlaylistModel)
-                ? (playlists[index]).getMap
+                ? ((playlists[index]) as PlaylistModel).getMap
                 : (playlists[index]);
 
             var playlistId = playlistMap["_id"];
@@ -178,7 +114,9 @@ class _LocalMusicPlaylistState extends State<LocalMusicPlaylist> {
             */
             return FutureBuilder<List<SongModel>>(
                 future: _audioQuery.queryAudiosFrom(
-                    AudiosFromType.PLAYLIST, playlistId),
+                  AudiosFromType.PLAYLIST,
+                  playlistId,
+                ),
                 builder: (ctx, i) {
                   if (i.hasError) return Text(i.error.toString());
                   // if (i.data == null) return const Center(child: CircularProgressIndicator());
@@ -199,10 +137,6 @@ class _LocalMusicPlaylistState extends State<LocalMusicPlaylist> {
                     playlist = playlists[index];
                   }
 
-                  print(
-                    '转为map之后再取值${playlist.runtimeType} $playlist-${playlist.playlist}',
-                  );
-
                   return ListTile(
                     selected: selectedPlaylists
                         .where((e) => e.id == playlist.id)
@@ -222,7 +156,13 @@ class _LocalMusicPlaylistState extends State<LocalMusicPlaylist> {
                       artworkHeight: 100.sp, // 这个高度显示不太对，实测始终是56，原因不明
                       artworkFit: BoxFit.cover,
                       keepOldArtwork: true, // 在生命周期内使用旧的缩略图
-                      nullArtworkWidget: const SizedBox.shrink(),
+                      // nullArtworkWidget: const SizedBox.shrink(),
+                      // 没有缩略图时使用占位图
+                      nullArtworkWidget: SizedBox(
+                        width: 100.sp,
+                        child: Image.asset(placeholderImageUrl,
+                            fit: BoxFit.fitWidth),
+                      ),
                     ),
 
                     onLongPress: () {
@@ -234,6 +174,7 @@ class _LocalMusicPlaylistState extends State<LocalMusicPlaylist> {
                       });
                     },
                     onTap: () {
+                      /// 如果已经处于歌单长按状态，点击则是支持以下功能：
                       if (llp.isPlaylistLongPress == LongPressStats.YES) {
                         setState(() {
                           // 如果已经加入被选中列表，再次点击则移除
@@ -247,13 +188,10 @@ class _LocalMusicPlaylistState extends State<LocalMusicPlaylist> {
                                 .removeWhere((e) => e.id == playlist.id);
                           } else {
                             selectedPlaylists.add(playlist);
-                            print(
-                              "selectedPlaylists===============$selectedPlaylists",
-                            );
                           }
+
                           // 如果被选中的列表清空，那就假装没有点击长按用于选择音频
                           if (selectedPlaylists.isEmpty) {
-                            print("8888888加载页面置为空？");
                             llp.changeIsPlaylistLongPress(LongPressStats.INIT);
                           }
 
@@ -261,32 +199,31 @@ class _LocalMusicPlaylistState extends State<LocalMusicPlaylist> {
                           llp.changeSelectedPlaylists(selectedPlaylists);
                         });
                       } else {
-                        print(
-                          '指定歌单 ${playlist.playlist}  was tapped! id Is ${playlist.id}',
-                        );
-
+                        /// 如果歌单不是处于长按状态，点击则是进入歌单列表页面：
                         Navigator.of(ctx)
-                            .push(
-                              MaterialPageRoute(
+                            .push(MaterialPageRoute(
                                 // 在选中指定歌单点击后，进入音频列表，同时监控是否有对音频长按
                                 builder: (BuildContext ctx) => MultiProvider(
-                                  providers: [
-                                    ListenableProvider<AudioLongPress>(
-                                      create: (_) => AudioLongPress(),
-                                    ),
-                                    ListenableProvider<AudioOptionSelected>(
-                                      create: (_) => AudioOptionSelected(),
-                                    ),
-                                  ],
-                                  child: LocalMusicAudioListDetail(
-                                    audioListType: AudioListTypes.playlist,
-                                    audioListId: playlist.id,
-                                    audioListTitle: playlist.playlist,
-                                  ),
-                                ),
-                              ),
-                            )
-                            .then((value) => initData());
+                                      providers: [
+                                        ListenableProvider<AudioLongPress>(
+                                          create: (_) => AudioLongPress(),
+                                        ),
+                                        ListenableProvider<AudioOptionSelected>(
+                                          create: (_) => AudioOptionSelected(),
+                                        ),
+                                      ],
+                                      child: LocalMusicAudioListDetail(
+                                        audioListType: AudioListTypes.playlist,
+                                        audioListId: playlist.id,
+                                        audioListTitle: playlist.playlist,
+                                      ),
+                                    )))
+                            .then((value) {
+                          // 进入指定歌单查看音频之后返回，需要重新查询歌单列表，以防有音频的删除信息而歌单列表显示未更新的问题
+                          setState(() {
+                            futureHandler = _audioQuery.queryPlaylists();
+                          });
+                        });
                       }
                     },
                   );
