@@ -116,7 +116,7 @@ class LabeledRadio extends StatelessWidget {
                 onChanged(newValue!);
               },
             ),
-            Text(label),
+            Text(label, style: TextStyle(fontSize: 13.sp)),
           ],
         ),
       ),
